TSP fitness is now 1/fitness - executes, but results doesnt seem correct.
I realised that the fitness is added up, thus higher fitness value =
better individual. For the TSP however it is the other way. The higher
the fitness, the longer the path to go. As long as a double float is
precise enough to hold fractions of 1/thousands it should work.